backend/requirements.txt
fastapi
uvicorn
pillow
torch
onnxruntime
python-multipart
pydantic
rich
